Removed the "evio" directory after moving the necessary files to "decoder_maindaq".
